backend/requirements.txt
# ✅ Compatible with Python 3.11 — all pre-built wheels
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
pdfplumber==0.10.3
PyPDF2==3.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.3.0
pydantic-settings==2.0.3
aiofiles==23.2.1
anyio==3.7.1
